refactor: split up flat persistent db and mem db in preparation of feature flags
